Table of skin color probabilities.
This code contains a program to generate a table of probabilities to be
used to filter colors based on its UV components.  The table is
generated from normal distributions, obtained from mean and variance of
selected colors in each component.  In this version, those parameters
are empirical.  To take the probability of a pixel, the table must be
indexed by values of components U and V.