Crops a human face in a picture
Using a model for the human skin color, human faces a filtered in a
image, saving space in disk when only human faces are insteresting for
the application.

The high-level algorithm is: filter pixels by color, reduce the image
scale, removing some remaining unisteresting pixels and get the part of
the picture that still has some interesting (non-dark) pixels.

The current model is a stub, to be a replaced by a model created using
some parameters. Some low-level implementations can be made more
accurate too.