Architecture change, rename and split
The file was renamed for better meaning of its functions. One of them,
was splitted in other file. This will break the build, that will fixed
in the following commits.

Image format is converted to a native format for processing, so some
functions that previously did saves from specific formats, now do
conversions of images.

In hcvloop file, a new function was created to enable a pipe of image
processing, that will use options from command line.