Creates another function to convert, process and save the image.
While the original process_image did the image format conversion, this
was suboptimal, since we could end up freeing an image that did not need
conversion and was not copied.

Taking code out of the loop in process_queue and putting it in a
function also makes things more readable. And this function may be
easily extended to allow us to save the full image even when we get a
face.