Using recreate flag to redraw on width change
We used to store the desired width and compare it to the width set to
see if we want to redraw. We can use the flag recreate to redraw
every time the width is set, and then we can check only once for this
flag and create the new surface.