Use a task to push buffers and repeat last buffer.
We use a task to push buffers to the source pad poping them from a
queue. The chain function pushes buffers to the queue. When no buffer is
available, push the last buffer. The task must be paused so we can
release the pad stream lock that is required to call the chain function.